Try separating FIPS and non-FIPS md5 tests
